requirements.txt
absl-py==1.1.0
addict==2.4.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
bce-python-sdk==0.8.64
beautifulsoup4==4.10.0
bleach==4.1.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
Cython==0.29.28
cython-bbox==0.1.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
Django==3.2.12
easydict==1.9
editdistance==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
fast_ctc_decode==0.3.0
fastseg==0.1.2
filelock==3.4.2
flake8==3.9.2
flake8-import-order==0.18.1
Flask==2.0.2
Flask-Babel==2.0.0
flatbuffers==1.12
fonttools==4.29.1
future==0.18.2
fvcore==0.1.5.post20220305
gast==0.3.3
gdown==4.2.1
geffnet==0.9.8
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
huggingface-hub==0.4.0
hyperopt==0.1.2
identify==2.4.7
idna==3.3
imagecodecs==2021.11.20
imageio==2.9.0
imageio-ffmpeg==0.4.5
imagezmq==1.1.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
imutils==0.5.4
iniconfig==1.1.1
iopath==0.1.9
ipyfilechooser==0.6.0
ipykernel==6.9.1
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
itsdangerous==2.0.1
jedi==0.18.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jstyleson==0.0.2
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lap==0.4.0
libclang==14.0.1
librosa==0.8.1
llvmlite==0.36.0
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mccabe==0.6.1
mediapipe==0.8.10
mistune==0.8.4
motmetrics==1.2.0
mpmath==1.2.1
multi-object-tracker==0.0.1
multiprocess==0.70.12.2
munch==2.5.0
natsort==8.1.0
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.6.3
nibabel==3.2.1
ninja==1.10.2.3
nltk==3.6.7
nncf==2.1.0
nodeenv==1.6.0
notebook==6.4.8
numba==0.53.1
numpy==1.21.6
oauthlib==3.2.0
onnx==1.9.0
onnx-simplifier==0.3.7
onnx2pytorch==0.4.1
onnxoptimizer==0.2.6
onnxruntime==1.10.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
openvino==2021.4.2
openvino-dev==2021.4.2
openvino-telemetry==2022.1.1
opt-einsum==3.3.0
packaging==21.3
paddle2onnx==0.9.0
paddlehub==2.2.0
paddlenlp==2.0.8
paddlepaddle==2.1.3
pafy==0.5.5
pandas==1.1.5
pandocfilters==1.5.0
parasail==1.2.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pip==21.2.2
platformdirs==2.4.1
pluggy==1.0.0
pooch==1.6.0
portalocker==2.4.0
ppgan==2.1.0
pre-commit==2.17.0
progress==1.6
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.4
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.14.0
pyDeprecate==0.3.2
pydicom==2.2.2
pydot==1.4.2
pyflakes==2.3.1
Pygments==2.11.2
pymongo==4.0.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyshine==0.0.9
PySocks==1.7.1
pytest==7.1.1
pytest-benchmark==3.4.1
python-dateutil==2.8.2
pytube==11.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
rarfile==4.0
rawpy==0.17.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
sacremoses==0.0.47
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools==58.0.4
Shapely==1.8.0
shellcheck-py==0.8.0.3
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
sqlparse==0.4.2
sympy==1.10.1
tabulate==0.8.9
tb-nightly==2.9.0a20220323
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow-datasets==4.2.0
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.6.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
tf2onnx==1.11.1
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.11.4
toml==0.10.2
tomli==2.0.1
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typing-extensions==3.7.4.3
urllib3==1.26.8
virtualenv==20.13.0
visualdl==2.2.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xmltodict==0.12.0
yacs==0.1.8
yamlloader==1.1.0
yapf==0.32.0
yaspin==2.1.0
youtube-dl==2020.12.2
zipp==3.7.0
zmq==0.0.0
